refactor to add model data to git_info
