+++ g3/git/git.py
@@ -1,8 +1,16 @@
-from typing import Optional
+from typing import List, Optional
 
 from pydantic import BaseModel
 
+from g3.git.diff import Diff, get_filenames, get_files_changed
 from g3.git.shell import Shell
+
+
+class ModelData(BaseModel):
+    filenames: List[str]
+    diffs: List[Diff]
+    repo_name: str
+    branch_name: str
 
 
 class GitInfo(BaseModel):
@@ -22,3 +30,14 @@ class GitInfo(BaseModel):
         repo_info = sh.git("config", "--get", "remote.origin.url").split("/")
         self.repo = repo_info[-1]
         self.repo_owner = repo_info[-2]
+
+    def get_data_for_model(self) -> ModelData:
+        sh = Shell()
+        assert sh.is_git()
+
+        return ModelData(
+            filenames=get_filenames(sh),
+            diffs=[diff for diff in get_files_changed(sh)],
+            repo_name=sh.repo_name,
+            branch_name=sh.branch_name,
+        )
