+++ tests/git/test_diff.py
@@ -1,3 +1,5 @@
+import os
+
 from g3.git.diff import exclude_files
 
 
@@ -8,4 +10,18 @@ def test_exclude_files():
 
 
 def test_get_files_changed():
-    pass
+    from g3.git.git import GitInfo
+    from g3.git.shell import Shell
+
+    git_info = GitInfo()
+    sh = Shell()
+
+    # Create temporary file in this directory and add it to the git index
+    with open("test.txt", "w") as f:
+        f.write("hello world")
+
+    sh.git("add", "test.txt")
+    assert "test.txt" in git_info.get_data_for_model().filenames
+
+    sh.git("rm", "--cached", "test.txt")
+    os.remove("test.txt")
